pyproject.toml
[project]
name = "product-price"
version = "0.1.0"
description = "Product Price Tracker"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydoll-python>=2.22.0",
    "tldextract>=5.3.1",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "ruff>=0.15.9",
]
